+++ src/app/(home)/pricing/page.tsx
@@ -21,7 +21,7 @@ const Page = () => {
           <div className="flex flex-col items-center">
             <Image
               src="/logo.svg"
-              alt="Sparal"
+              alt="LaunchKit AI"
               width={50}
               height={50}
               className="hidden md:block"